requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alabaster==0.7.13
arrow==1.3.0
astroid==3.2.4
asttokens==3.0.0
async-timeout==5.0.1
asyncssh==2.20.0
atomicwrites==1.4.1
attrs==25.3.0
autopep8==2.0.4
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
black==24.8.0
bleach==6.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
cookiecutter==2.6.0
cryptography==44.0.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
dill==0.3.9
docstring-to-markdown==0.15
docutils==0.20.1
executing==2.2.0
Faker==35.2.2
fastjsonschema==2.21.1
flake8==7.1.2
frozenlist==1.5.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.12.3
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.5.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.24.4
numpydoc==1.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.9.0
pylint==3.2.7
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.5.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.9.0.post0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.1
pyuca==1.2
pywin32==309
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.3.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rope==1.13.0
rpds-py==0.20.1
rsa==4.9
Rtree==1.3.0
six==1.17.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==6.0.4
spyder-kernels==3.0.3
stack-data==0.6.3
superqt==0.6.7
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.3
three-merge==0.1.1
tinycss2==1.2.1
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
whatthepatch==1.0.6
wrapt==1.17.2
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
